requirements.txt
absl-py==0.12.0 
aiohttp
astor==0.8.1
astunparse==1.6.3
async-timeout==3.0.1
attrs==21.2.0 
blinker==1.4
brotlipy==0.7.0
CacheControl==0.12.6
cachetools==4.2.2
certifi==2020.12.5
cffi==1.14.5 
chardet
click==7.1.2
cryptography 
firebase-admin==5.0.0
Flask==1.1.2
gast
gevent==21.1.2
glob2==0.7
google-api-core==1.28.0
google-api-python-client==2.6.0
google-auth==1.30.0 
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.4.1
google-cloud-core==1.6.0
google-cloud-firestore==2.1.1
google-cloud-storage==1.38.0
google-crc32c==1.1.2
google-pasta==0.2.0
google-resumable-media==1.3.0
googleapis-common-protos==1.53.0
greenlet==1.1.0
grpcio==1.38.0 
h5py==2.10.0
httplib2==0.19.1
idna==2.10 
importlib-metadata==4.0.1 
itsdangerous==1.1.0
Jinja2==2.11.2
Keras==2.4.3 
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
Markdown==3.3.4
MarkupSafe==1.1.1
msgpack==1.0.2
multidict==5.1.0 
numpy
oauthlib==3.0.1
opt-einsum==3.3.0
opencv-python 
packaging==20.9
Pillow==8.2.0
proto-plus==1.18.1
protobuf==3.17.0
pyasn1==0.4.8
pyasn1-modules==0.2.7
pycparser==2.20 
pyfcm==1.5.1
PyJWT==2.1.0
pyOpenSSL==20.0.1
pyparsing==2.4.7
pyreadline==2.1 
PySocks==1.7.1 
pytz==2021.1
PyYAML==5.4.1
requests==2.25.1
requests-oauthlib==1.3.0 
rsa==4.7.2 
scipy
six==1.16.0 
tensorboard==2.4.1 
tensorboard-plugin-wit==1.8.0 
tensorflow==2.3.0
tensorflow-estimator
termcolor==1.1.0
typing-extensions==3.7.4.3
uritemplate==3.0.1
urllib3==1.26.4 
Werkzeug==1.0.1
win-inet-pton==1.1.0 
wrapt==1.12.1 
yarl==1.6.3 
zipp==3.4.1 
zope.event==4.5.0
zope.interface==5.4.0